+++ src/strategies/multimodal_strategy.py
@@ -285,7 +285,7 @@ def _run_epoch(model, loader, optimizer, scaler, device, config, weights, *, tra
                 scaler.scale(loss).backward()
                 scaler.step(optimizer)
                 scaler.update()
-            probability = logits.detach().softmax(1)
+            probability = logits.detach().float().softmax(1)
             total_loss += float(loss.detach()) * len(target)
             targets.extend(target.cpu().tolist())
             probabilities.extend(probability.cpu().tolist())
